complete the change of flow when the document is not a contract
